@@ -1,12 +1,12 @@
 buildscript {
-    ext.kotlin_version = '1.2.30'
+    ext.kotlin_version = '1.2.61'
     repositories {
         mavenCentral()
         maven { url 'https://oss.sonatype.org/content/repositories/snapshots/' }
         google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.1.2'
+        classpath 'com.android.tools.build:gradle:3.2.0-rc02'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
     }
 }
@@ -26,24 +26,27 @@ apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-kapt'
 
 dependencies {
-    implementation 'com.android.support:support-v13:27.1.1'
-    implementation 'com.android.support:support-v4:27.1.1'
-    implementation 'com.android.support:cardview-v7:27.1.1'
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support:design:27.1.1'
-    implementation 'com.android.support:preference-v7:27.1.1'
-    implementation 'com.android.support:palette-v7:27.1.1'
+    implementation 'androidx.cardview:cardview:1.0.0-rc02'
+    implementation 'androidx.appcompat:appcompat:1.0.0-rc02'
+    implementation 'com.google.android.material:material:1.0.0-rc01'
+    implementation 'androidx.preference:preference:1.0.0-rc02'
+    implementation 'androidx.palette:palette:1.0.0-rc02'
 
     implementation 'com.squareup.picasso:picasso:2.71828'
-    implementation 'com.android.support:exifinterface:27.1.1'
+    implementation 'androidx.exifinterface:exifinterface:1.0.0-rc02'
 
     // Activity recognition
     implementation 'com.google.android.gms:play-services-location:15.0.1'
 
-    implementation 'androidx.core:core-ktx:0.3'
+    implementation 'androidx.core:core-ktx:1.0.0-rc02'
+    implementation 'androidx.fragment:fragment-ktx:1.0.0-rc02'
+    implementation 'androidx.collection:collection-ktx:1.0.0-rc02'
+    implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.0.0-rc01'
+    implementation 'androidx.palette:palette-ktx:1.0.0-rc02'
+    implementation 'androidx.sqlite:sqlite-ktx:2.0.0-rc01'
 
-    implementation 'android.arch.lifecycle:extensions:1.1.1'
-    kapt 'android.arch.lifecycle:compiler:1.1.1'
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.0.0-rc01'
+    kapt 'androidx.lifecycle:lifecycle-compiler:2.0.0-rc01'
 
     implementation "ch.acra:acra-core:5.1.1"
     implementation "ch.acra:acra-dialog:5.1.1"
@@ -51,14 +54,14 @@ dependencies {
 }
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 28
+    buildToolsVersion '28.0.2'
 
     defaultConfig {
         minSdkVersion 21
-        targetSdkVersion 25
-        versionCode 68
-        versionName '1.9.4'
+        targetSdkVersion 28
+        versionCode 200
+        versionName '2.0.0'
     }
 
     signingConfigs {
