@@ -2,7 +2,7 @@ package com.anod.car.home
 
 import android.app.Application
 import android.content.Context
-import android.support.v7.app.AppCompatDelegate
+import androidx.appcompat.app.AppCompatDelegate
 
 import com.anod.car.home.acra.BrowserUrlSender
 import com.anod.car.home.acra.CrashDialog
