@@ -8,7 +8,7 @@ import org.acra.prefs.SharedPreferencesFactory
 
 import android.content.DialogInterface
 import android.os.Bundle
-import android.support.v7.app.AlertDialog
+import androidx.appcompat.app.AlertDialog
 import android.view.LayoutInflater
 import android.view.View
 import android.widget.EditText
