@@ -2,7 +2,7 @@ package com.anod.car.home.appwidget;
 
 import android.app.PendingIntent;
 import android.content.Context;
-import android.support.annotation.IdRes;
+import androidx.annotation.IdRes;
 import android.view.View;
 import android.widget.RemoteViews;
 
