@@ -1,7 +1,7 @@
 package com.anod.car.home.prefs.model;
 
 import android.content.ComponentName;
-import android.support.v4.util.ArrayMap;
+import androidx.collection.ArrayMap;
 
 /**
  * @author algavris
