@@ -3,8 +3,8 @@ package com.anod.car.home.prefs.model;
 import android.content.SharedPreferences;
 import android.content.res.Resources;
 import android.graphics.Color;
-import android.support.v4.content.res.ResourcesCompat;
-import android.support.v4.util.SimpleArrayMap;
+import androidx.core.content.res.ResourcesCompat;
+import androidx.collection.SimpleArrayMap;
 import android.util.JsonReader;
 import android.util.JsonToken;
 import android.util.JsonWriter;
