@@ -3,7 +3,7 @@ package com.anod.car.home.prefs.views.drag;
 import android.annotation.SuppressLint;
 import android.graphics.Canvas;
 import android.graphics.Point;
-import android.support.v4.content.res.ResourcesCompat;
+import androidx.core.content.res.ResourcesCompat;
 import android.view.View;
 
 import com.anod.car.home.R;
