@@ -2,7 +2,7 @@ package com.anod.car.home.utils;
 
 import com.anod.car.home.R;
 
-import android.support.v7.app.AlertDialog;
+import androidx.appcompat.app.AlertDialog;
 import android.app.Dialog;
 import android.content.Context;
 import android.content.DialogInterface;
